fix: replace the gatsby favicon and update the manifest settings
This was driving me absolutely bonkers. Every time I saw one of your articles pop up on my web browser, it was with the gatsby favicon and the wrong name. This at maximum fixes the problem and at minimum alerts you to it. I also compressed the other images that were in the same folder.